+++ react/utils/src/atom/useEasyGetAtom.ts
@@ -1,5 +1,5 @@
 import { selectAtom } from '@einfach/core'
-import type { HookOption } from '@einfach/react';
+import type { HookOption } from '@einfach/react'
 import { useAtomValue } from '@einfach/react'
 import type { AtomEntity } from '@einfach/core'
 import { easyGet, easyEqual } from '@einfach/utils'
@@ -25,10 +25,10 @@ export function useEasySelectAtomValue<Slice, State>(
   selectFn: NamePath,
   equalityFn: (prev: Slice, next: Slice) => boolean = easyEqual,
   option: HookOption = {},
-) {
+): Slice {
   const selectAtomEntity = useInit(() => {
     return selectEasyAtom<Slice, State>(atomEntity, selectFn, equalityFn)
   }, [atomEntity, selectFn, equalityFn])
 
-  return useAtomValue<Slice>(selectAtomEntity, option)
+  return useAtomValue(selectAtomEntity, option) as Slice
 }