+++ solid/solid/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @einfach/solid
 
+## 0.1.4
+
+### Patch Changes
+
+- 更新selectATtom的类型
+- Updated dependencies
+  - @einfach/core@0.1.6
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.1.5
+
 ## 0.1.2
 
 ### Patch Changes